Add Met Office locations
This JSON file should eventually be periodically fetched and cached from 
the Meto Office website. However, it will not change very often, so there is 
no need to do so unless this service starts being used.

This is in a 'lib' directory as the javascript to find a location was provided by 
the Met Office, with some minor adaptations to enable it to be imported 
into ES6 modules.