Add an API for finding locations from names
Uses the levenshtein distance to order the closest matches first.
The Ramda library is used to enable a functional style. Node's
require function is used to get around the levenshtein library
not having Typescript definitions and to require the JSON locations.
In addition, the met office lib files are now ES5 so that they do
not need to be compiled, which is unnecessary as the code is just
ES5.