Add Met Office API
This introduces the concept of 'Sources' which are sources of weather
data. They are broken up into several parts - fetching (asynchronous),
transforming (synchronous) and interface definition. Their role is to
fetch data from an external source and transform the returned data
from the external format into a well-defined single format (ForecastPoint).
In addition, each source should have an id (for API URLs), name and
URL so that it can be linked to and have its contribution attributed.